got the system fully working, though telemetry is still buggy, likely due to being too fast for the logger
