+++ src/main/java/frc/robot/subsystems/swervedrive/SwerveDriveSubsystem.java
@@ -1,109 +1,116 @@
-import edu.wpi.first.math.Matrix;
-import edu.wpi.first.math.VecBuilder;
-import edu.wpi.first.math.controller.ProfiledPIDController;
-import edu.wpi.first.math.estimator.SwerveDrivePoseEstimator;
-import edu.wpi.first.math.filter.LinearFilter;
-import edu.wpi.first.math.geometry.Pose2d;
-import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.geometry.Rotation3d;
-import edu.wpi.first.math.geometry.Translation2d;
-import edu.wpi.first.math.geometry.Translation3d;
-import edu.wpi.first.math.geometry.proto.Pose2dProto;
-import edu.wpi.first.math.kinematics.ChassisSpeeds;
-import edu.wpi.first.math.kinematics.SwerveDriveKinematics;
-import edu.wpi.first.math.kinematics.SwerveModulePosition;
-import edu.wpi.first.math.kinematics.SwerveModuleState;
-import edu.wpi.first.math.numbers.N1;
-import edu.wpi.first.math.numbers.N3;
-import edu.wpi.first.math.proto.Geometry2D.ProtobufPose2d;
-import edu.wpi.first.math.trajectory.TrapezoidProfile;
-import edu.wpi.first.wpilibj.DriverStation;
-import edu.wpi.first.wpilibj.Timer;
-import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.Commands;
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.lib.interpolation.MovingAverageVelocity;
-import frc.lib.logging.LoggedReceiver;
-import frc.lib.logging.Logger;
-import frc.lib.math.MathUtils;
-import frc.lib.swerve.SwerveDriveSignal;
-import frc.robot.Constants;
-import frc.robot.Constants.SwerveConstants;
-import frc.robot.commands.DriveToPositionCommand;
-import frc.robot.subsystems.swervedrive.GyroIO.GyroIOInputs;
+package frc.robot.subsystems.swervedrive;
 
-import java.util.List;
 import java.util.function.DoubleSupplier;
+import java.util.function.Supplier;
 
+import com.ctre.phoenix6.Utils;
+import com.ctre.phoenix6.mechanisms.swerve.SwerveDrivetrain;
+import com.ctre.phoenix6.mechanisms.swerve.SwerveDrivetrainConstants;
+import com.ctre.phoenix6.mechanisms.swerve.SwerveModuleConstants;
+import com.ctre.phoenix6.mechanisms.swerve.SwerveRequest;
+import com.ctre.phoenix6.mechanisms.swerve.SwerveModule.DriveRequestType;
+import com.ctre.phoenix6.mechanisms.swerve.SwerveModule.SteerRequestType;
 import com.pathplanner.lib.auto.AutoBuilder;
-import com.pathplanner.lib.path.GoalEndState;
 import com.pathplanner.lib.path.PathConstraints;
-import com.pathplanner.lib.path.PathPlannerPath;
-import com.pathplanner.lib.pathfinding.Pathfinding;
 import com.pathplanner.lib.util.HolonomicPathFollowerConfig;
 import com.pathplanner.lib.util.PIDConstants;
 import com.pathplanner.lib.util.ReplanningConfig;
 
-public class SwerveDriveSubsystem extends CommandSwerveDrivetrain {
-    private final SwerveDrivePoseEstimator swervePoseEstimator;
+import edu.wpi.first.math.controller.ProfiledPIDController;
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.geometry.Rotation3d;
+import edu.wpi.first.math.kinematics.ChassisSpeeds;
+import edu.wpi.first.math.trajectory.TrapezoidProfile;
+import edu.wpi.first.wpilibj.DriverStation;
+import edu.wpi.first.wpilibj.Notifier;
+import edu.wpi.first.wpilibj.RobotController;
+import edu.wpi.first.wpilibj2.command.Command;
+import edu.wpi.first.wpilibj2.command.Commands;
+import edu.wpi.first.wpilibj2.command.Subsystem;
+import frc.lib.logging.Logger;
+import frc.lib.math.MathUtils;
+import frc.robot.Constants;
+import frc.robot.Telemetry;
+import frc.robot.TunerConstants;
+import frc.robot.commands.DriveToPositionCommand;
 
-    private Pose2d pose = new Pose2d();
-    private final MovingAverageVelocity velocityEstimator = new MovingAverageVelocity(3);
-    private ChassisSpeeds velocity = new ChassisSpeeds(0,0,0);
-    private ChassisSpeeds previousVelocity = new ChassisSpeeds();
-    private SwerveDriveSignal driveSignal = new SwerveDriveSignal();
-    private ChassisSpeeds acceleration = new ChassisSpeeds(0,0,0);
-    private LinearFilter accelerationX = LinearFilter.movingAverage(5);
-    private LinearFilter accelerationY = LinearFilter.movingAverage(5);
-    private LinearFilter accelerationOmega = LinearFilter.movingAverage(5);
+/**
+ * Class that extends the Phoenix SwerveDrivetrain class and implements subsystem
+ * so it can be used in command-based projects easily.
+ */
+public class SwerveDriveSubsystem extends SwerveDrivetrain implements Subsystem {
+    private static final double kSimLoopPeriod = 0.005; // 5 ms
+    private Notifier m_simNotifier = null;
+    private double m_lastSimTime;
 
-    public boolean isRainbow = false;
+    public final SwerveRequest.RobotCentric closedLoopRobotCentric = new SwerveRequest.RobotCentric()
+                    .withDriveRequestType(DriveRequestType.Velocity)
+                    .withSteerRequestType(SteerRequestType.MotionMagicExpo);
 
-    private SwerveModule[] modules;
+    public final SwerveRequest.FieldCentric closedLoop = new SwerveRequest.FieldCentric()
+                    .withDriveRequestType(DriveRequestType.Velocity)
+                    .withSteerRequestType(SteerRequestType.MotionMagicExpo);
 
-    private CommandSwerveDrivetrain drivetrain;
+    public final SwerveRequest.FieldCentric openLoop = new SwerveRequest.FieldCentric()
+                    .withDriveRequestType(DriveRequestType.OpenLoopVoltage)
+                    .withSteerRequestType(SteerRequestType.MotionMagicExpo);
 
-    private GyroIO gyroIO;
-    private GyroIOInputs gyroInputs = new GyroIOInputs();
+    public final SwerveRequest.ApplyChassisSpeeds stopped = new SwerveRequest.ApplyChassisSpeeds()
+                    .withDriveRequestType(DriveRequestType.OpenLoopVoltage)
+                    .withSteerRequestType(SteerRequestType.MotionMagicExpo);
 
-    boolean isCharacterizing = false;
+    public SwerveDriveSubsystem(SwerveDrivetrainConstants driveTrainConstants, double OdometryUpdateFrequency, SwerveModuleConstants... modules) {
+        super(driveTrainConstants, OdometryUpdateFrequency, modules);
+        applyConfigs();
+        if (Utils.isSimulation()) {
+            startSimThread();
+        }
+    }
 
-    private LoggedReceiver isSecondOrder;
+    public SwerveDriveSubsystem(SwerveDrivetrainConstants driveTrainConstants, SwerveModuleConstants... modules) {
+        super(driveTrainConstants, modules);
+        applyConfigs();
+        if (Utils.isSimulation()) {
+            startSimThread();
+        }
+    }
 
-    // Cardinal command
-    private ProfiledPIDController omegaController =
-            new ProfiledPIDController(5, 0, 0, new TrapezoidProfile.Constraints(8, 8));
-    private final double maxCardinalVelocity = 4.5;
+    public ChassisSpeeds getRobotRelativeChassisSpeeds() {
+        return m_kinematics.toChassisSpeeds(getState().ModuleStates);
+    }
 
-    private final double angularVelocityCoefficient = 0.0;
-    private final double accelerationCoefficient = 0; //0.12;
+    public ChassisSpeeds getFieldRelativeChassisSpeeds() {
+        return ChassisSpeeds.fromRobotRelativeSpeeds(getRobotRelativeChassisSpeeds(), getRotation());
+    }
 
-    private ChassisSpeeds lastDriveSignalClosedLoop = new ChassisSpeeds(0,0,0);
-    private double lastDriveSignalClosedLoopTimestamp = 0;
+    public double getVelocityMagnitude() {
+        ChassisSpeeds currentSpeeds = getRobotRelativeChassisSpeeds();
+        return Math.hypot(currentSpeeds.vxMetersPerSecond, currentSpeeds.vyMetersPerSecond);
+    }
 
-    private double previousTilt = 0;
-    private double tiltRate = 0;
-    private DoubleSupplier maxSpeedSupplier = () -> Constants.SwerveConstants.maxSpeed;
+    public Pose2d getPose() {
+        return getState().Pose;
+    }
 
-    public SwerveDriveSubsystem(CommandSwerveDrivetrain drivetrain) {
-        
-        // Initialize the swerve drive pose estimator with access to the module positions.
-        swervePoseEstimator = new SwerveDrivePoseEstimator(
-                SwerveConstants.swerveKinematics,
-                getGyroRotation(),
-                getModulePositions(),
-                new Pose2d(),
-                VecBuilder.fill(0.01, 0.01, 0.01),
-                VecBuilder.fill(0.3, 0.3, 0.9)); // might need to bring these back up (was 0.5)
+    public Rotation2d getRotation() {
+        return getPose().getRotation();
+    }
 
-        // Allow us to toggle on second order kinematics
-        isSecondOrder = Logger.tunable("/SwerveDriveSubsystem/isSecondOrder", false);
+    public Rotation2d getTilt() {
+        return new Rotation2d(Math.acos(getRotation3d().getAxis().get(2,0)));
+    }
 
-        AutoBuilder.configureHolonomic(
+    private void applyConfigs() {
+         AutoBuilder.configureHolonomic(
                 this::getPose, // Robot pose supplier
-                this::setPose, // Method to reset odometry (will be called if your auto has a starting pose)
-                this::getVelocity, // ChassisSpeeds supplier. MUST BE ROBOT RELATIVE
-                (velocity) -> setVelocityAccelerationPredict(velocity, false), // Method that will drive the robot given ROBOT RELATIVE ChassisSpeeds
+                this::seedFieldRelative, // Method to res.et odometry (will be called if your auto has a starting pose)
+                this::getRobotRelativeChassisSpeeds, // ChassisSpeeds supplier. MUST BE ROBOT RELATIVE
+                (velocity) -> this.setControl(
+                    closedLoopRobotCentric
+                    .withVelocityX(velocity.vxMetersPerSecond)
+                    .withVelocityY(velocity.vyMetersPerSecond)
+                    .withRotationalRate(velocity.omegaRadiansPerSecond)), // Method that will drive the robot given ROBOT RELATIVE ChassisSpeeds
                 new HolonomicPathFollowerConfig( // HolonomicPathFollowerConfig, this should likely live in your Constants class
                         new PIDConstants(5, 0.0, 0.01), // Translation PID constants
                         new PIDConstants(4, 0.0, 0.01), // Rotation PID constants
@@ -124,15 +131,51 @@ public class SwerveDriveSubsystem extends CommandSwerveDrivetrain {
                 },
                 this // Reference to this subsystem to set requirements
         );
+
+        Telemetry myTelem = new Telemetry(TunerConstants.kSpeedAt12VoltsMps);
+
+        registerTelemetry((state) -> myTelem.telemeterize(state));
+    }
+
+    public Command applyRequest(Supplier<SwerveRequest> requestSupplier) {
+        return run(() -> this.setControl(requestSupplier.get()));
     }
 
     public Command driveCommand(
-            DoubleSupplier forward, DoubleSupplier strafe, DoubleSupplier rotation, boolean isFieldOriented) {
+            DoubleSupplier forward, DoubleSupplier strafe, DoubleSupplier rotation) {
+        return applyRequest(
+                    () ->openLoop.withDeadband(0.0).withRotationalDeadband(0.0)
+                    .withVelocityX(forward.getAsDouble())
+                    .withVelocityY(strafe.getAsDouble())
+                    .withRotationalRate(rotation.getAsDouble())
+        );
+    }
+
+    public Command cardinalCommand(Rotation2d targetAngle, DoubleSupplier forward, DoubleSupplier strafe) {
+        final ProfiledPIDController omegaController =
+            new ProfiledPIDController(5, 0, 0, new TrapezoidProfile.Constraints(8, 8));
+
+        final double maxCardinalVelocity = 3.0;
+
+        omegaController.enableContinuousInput(-Math.PI, Math.PI);   
+
         return run(() -> {
-            setVelocity(
-                    new ChassisSpeeds(forward.getAsDouble(), strafe.getAsDouble(), rotation.getAsDouble()),
-                    isFieldOriented);
-        });
+                    var rotationCorrection =
+                            omegaController.calculate(getPose().getRotation().getRadians(), targetAngle.getRadians());
+
+                    setControl(
+                            openLoop
+                                .withVelocityX(forward.getAsDouble())
+                                .withVelocityY(strafe.getAsDouble())
+                                .withRotationalRate(
+                                    MathUtils.ensureRange(
+                                            rotationCorrection, -maxCardinalVelocity, maxCardinalVelocity)
+                            ));
+                })
+                .beforeStarting(() -> {
+                    omegaController.reset(new TrapezoidProfile.State(
+                            getRotation().getRadians(), getFieldRelativeChassisSpeeds().omegaRadiansPerSecond));
+                });
     }
 
     public Command pathfindToPoseCommand(Pose2d targetPose) {
@@ -158,320 +201,65 @@ public class SwerveDriveSubsystem extends CommandSwerveDrivetrain {
         return new DriveToPositionCommand(this, targetPose);
     }
 
-    public Command preciseDriveCommand(
-            DoubleSupplier forward, DoubleSupplier strafe, DoubleSupplier rotation, boolean isFieldOriented) {
-        var speedMultiplier = SwerveConstants.preciseDrivingModeSpeedMultiplier;
+    private void startSimThread() {
+        m_lastSimTime = Utils.getCurrentTimeSeconds();
 
-        return run(() -> {
-            setVelocity(
-                    new ChassisSpeeds(
-                            speedMultiplier * forward.getAsDouble(),
-                            speedMultiplier * strafe.getAsDouble(),
-                            speedMultiplier * rotation.getAsDouble()),
-                    isFieldOriented);
+        /* Run simulation at a faster rate so PID gains behave more reasonably */
+        m_simNotifier = new Notifier(() -> {
+            final double currentTime = Utils.getCurrentTimeSeconds();
+            double deltaTime = currentTime - m_lastSimTime;
+            m_lastSimTime = currentTime;
+
+            /* use the measured time delta, get battery voltage from WPILib */
+            updateSimState(deltaTime, RobotController.getBatteryVoltage());
         });
+        m_simNotifier.startPeriodic(kSimLoopPeriod);
     }
 
-    public Command cardinalCommand(Rotation2d targetAngle, DoubleSupplier forward, DoubleSupplier strafe) {
-        omegaController.enableContinuousInput(-Math.PI, Math.PI);
-
-        return run(() -> {
-                    var rotationCorrection =
-                            omegaController.calculate(pose.getRotation().getRadians(), targetAngle.getRadians());
-
-                    setVelocity(
-                            new ChassisSpeeds(
-                                    forward.getAsDouble(),
-                                    strafe.getAsDouble(),
-                                    MathUtils.ensureRange(
-                                            rotationCorrection, -maxCardinalVelocity, maxCardinalVelocity)),
-                            true);
-                })
-                .beforeStarting(() -> {
-                    omegaController.reset(new TrapezoidProfile.State(
-                            pose.getRotation().getRadians(), velocity.omegaRadiansPerSecond));
-                });
-    }
-
-    public void setCustomMaxSpeedSupplier(DoubleSupplier maxSpeedSupplier) {
-        this.maxSpeedSupplier = maxSpeedSupplier;
-    }
-
-    public Pose2d getPose() {
-        return pose;
-    }
-
-    public void setPose(Pose2d pose) {
-        this.pose = pose;
-        swervePoseEstimator.resetPosition(getGyroRotation(), getModulePositions(), pose);
-    }
-
-    public void addVisionPoseEstimate(Pose2d pose, double timestamp) {
-        CommandSwerveDrivetrain
-    }
-
-    public void addVisionPoseEstimate(Pose2d pose, double timestamp, Matrix<N3, N1> stdDevs) {
-        swervePoseEstimator.addVisionMeasurement(pose, timestamp, stdDevs);
-    }
-
-    /**
-     * @return The robot relative velocity of the drivetrain
-     */
-    public ChassisSpeeds getVelocity() {
-        return velocity;
-    }
-
-    public ChassisSpeeds getAcceleration() {
-        return new ChassisSpeeds(
-                velocity.vxMetersPerSecond - previousVelocity.vxMetersPerSecond,
-                velocity.vyMetersPerSecond - previousVelocity.vyMetersPerSecond,
-                velocity.omegaRadiansPerSecond - previousVelocity.omegaRadiansPerSecond);
-    }
-
-    /**
-     * @return The potentially field relative desired velocity of the drivetrain
-     */
-    public ChassisSpeeds getDesiredVelocity() {
-        return (ChassisSpeeds) driveSignal;
-    }
-
-    public double getVelocityMagnitude() {
-        return Math.sqrt(Math.pow(velocity.vxMetersPerSecond, 2) + Math.pow(velocity.vyMetersPerSecond, 2));
-    }
-
-    public Rotation2d getVelocityRotation() {
-        return (new Translation2d(velocity.vxMetersPerSecond, velocity.vxMetersPerSecond)).getAngle();
-    }
-
-    public ChassisSpeeds getSmoothedVelocity() {
-        return velocityEstimator.getAverage();
-    }
-
-    public Rotation2d getGyroRotation() {
-        return gyroInputs.rotation2d;
-    }
-
-    public Rotation2d getRotation() {
-        return pose.getRotation();
-    }
-
-    public Rotation3d getGyroRotation3d() {
-        return gyroInputs.rotation3d;
-    }
-
-    public Translation3d getNormalVector3d() {
-        return new Translation3d(0, 0, 1).rotateBy(getGyroRotation3d());
-    }
-
-    /**is in degrees*/
-    public double getTiltAmountInDegrees() {
-        return Math.toDegrees(getTiltAmount());
-    }
-
-    public double getTiltAmount() {
-        return Math.acos(getNormalVector3d().getZ());
-    }
-
-    public double getTiltRate() {
-        return tiltRate;
-    }
-
-    public Rotation2d getTiltDirection() {
-        return new Rotation2d(getNormalVector3d().getX(), getNormalVector3d().getY());
-    }
-
-    public void setRotation(Rotation2d angle) {
-        setPose(new Pose2d(getPose().getX(), getPose().getY(), angle));
-    }
-
-    public void zeroRotation() {
-        setRotation(new Rotation2d());
-    }
-
-    public void updateAcceleration(ChassisSpeeds velocity) {
-        double currentTime = Timer.getFPGATimestamp();
-        ChassisSpeeds temporalAcceleration = velocity.minus(lastDriveSignalClosedLoop).div(currentTime - lastDriveSignalClosedLoopTimestamp);
-        lastDriveSignalClosedLoop = velocity;
-        lastDriveSignalClosedLoopTimestamp = currentTime;
-        double x = accelerationX.calculate(temporalAcceleration.vxMetersPerSecond);
-        double y = accelerationY.calculate(temporalAcceleration.vyMetersPerSecond);
-        double omega = accelerationOmega.calculate(temporalAcceleration.omegaRadiansPerSecond);
-        acceleration = new ChassisSpeeds(x,y,omega);
-    }
-
-    public void setVelocityAcceleration(ChassisSpeeds velocity, ChassisSpeeds accelerationMeasured, boolean isFieldOriented) {
-        updateAcceleration(velocity);
-        driveSignal = new SwerveDriveSignal(velocity.plus(accelerationMeasured.times(accelerationCoefficient)), isFieldOriented, false);
-    }
-
-    public void setVelocityAccelerationPredict(ChassisSpeeds velocity, boolean isFieldOriented) {
-        updateAcceleration(velocity);
-        driveSignal = new SwerveDriveSignal(velocity.plus(acceleration.times(accelerationCoefficient)), isFieldOriented, false);
-    }
-
-    public void setVelocity(ChassisSpeeds velocity, boolean isFieldOriented, boolean isOpenLoop) {
-        updateAcceleration(velocity);
-        driveSignal = new SwerveDriveSignal(velocity, isFieldOriented, isOpenLoop);
-    }
-
-    public void setVelocity(ChassisSpeeds velocity, boolean isFieldOriented) {
-        setVelocity(velocity, isFieldOriented, true);
-    }
-
-    public void setVelocity(ChassisSpeeds velocity) {
-        setVelocity(velocity, false);
-    }
-
-    public void stop() {
-        driveSignal = new SwerveDriveSignal();
-    }
-
-    public void lock() {
-        driveSignal = new SwerveDriveSignal(true);
-    }
-
-    private void setModuleStates(SwerveModuleState[] desiredStates, boolean isOpenLoop) {
-        SwerveDriveKinematics.desaturateWheelSpeeds(desiredStates, maxSpeedSupplier.getAsDouble());
-
-        for (SwerveModule module : modules) {
-            module.setDesiredState(desiredStates[module.moduleNumber], isOpenLoop, isSecondOrder.getBoolean());
-        }
-
-        Logger.log("/SwerveDriveSubsystem/Wheel Setpoint", desiredStates[0].speedMetersPerSecond);
-    }
-
-    private boolean isDriveSignalStopped(SwerveDriveSignal driveSignal) {
-        return driveSignal.vxMetersPerSecond == 0
-                && driveSignal.vyMetersPerSecond == 0
-                && driveSignal.omegaRadiansPerSecond == 0;
-    }
-
-    @Override
-    public void periodic() {
-        gyroIO.updateInputs(gyroInputs);
-
-        for (SwerveModule module : modules) {
-            module.updateInputs();
-        }
-
-        var startTimeMS = Timer.getFPGATimestamp() * 1000;
-
-        Logger.log("/SwerveDriveSubsystem/LoopDuration", Timer.getFPGATimestamp() * 1000 - startTimeMS);
-
-        var tilt = getTiltAmount();
-        tiltRate = (tilt - previousTilt) / 0.02;
-        previousTilt = tilt;
-
-        updateOdometry();
-
-        if (isCharacterizing) return;
-
-        updateModules(driveSignal);
-        //* endd comment out here */
-
-        updateLogs();
-    }
-
-    private void updateOdometry() {
-        SwerveModuleState[] moduleStates = getModuleStates();
-        SwerveModulePosition[] modulePositions = getModulePositions();
-        
-        previousVelocity = velocity;
-        velocity = Constants.SwerveConstants.swerveKinematics.toChassisSpeeds(moduleStates);
-
-        // velocityEstimator.add(velocity);
-
-
-
-        if (gyroInputs.isActive) {
-            pose = swervePoseEstimator.update(getGyroRotation(), modulePositions);
-        } else { //TODO: this is a very hacky hack that should be fixed 
-            pose = swervePoseEstimator.update(pose.getRotation().plus(Rotation2d.fromRadians(velocity.omegaRadiansPerSecond * 0.05)), modulePositions);
-        }
-        
-    }
-
-    private void updateModules(SwerveDriveSignal driveSignal) {
-        ChassisSpeeds chassisVelocity;
-
-        if (driveSignal.isFieldOriented()) {
-            chassisVelocity = ChassisSpeeds.fromFieldRelativeSpeeds(
-                    driveSignal.vxMetersPerSecond,
-                    driveSignal.vyMetersPerSecond,
-                    driveSignal.omegaRadiansPerSecond,
-                    getRotation());
-        } else {
-            chassisVelocity = (ChassisSpeeds) driveSignal;
-        }
-
-        // 2910 twisting field correction
-        chassisVelocity = ChassisSpeeds.fromFieldRelativeSpeeds(chassisVelocity, new Rotation2d(chassisVelocity.omegaRadiansPerSecond * angularVelocityCoefficient));
-
-        SwerveModuleState[] moduleStates =
-                Constants.SwerveConstants.swerveKinematics.toSwerveModuleStates(chassisVelocity);
-
-        if (driveSignal.isLocked()) {
-            // get X for stopping
-            moduleStates = new SwerveModuleState[] {
-                new SwerveModuleState(0, Rotation2d.fromDegrees(45)),
-                new SwerveModuleState(0, Rotation2d.fromDegrees(-45)),
-                new SwerveModuleState(0, Rotation2d.fromDegrees(-45)),
-                new SwerveModuleState(0, Rotation2d.fromDegrees(45)),
-            };
-
-            // Set the angle of each module only
-            for (int i = 0; i < moduleStates.length; i++) {
-                modules[i].setDesiredAngleOnly(moduleStates[i].angle, true);
-            }
-        } else {
-            setModuleStates(moduleStates, isDriveSignalStopped(driveSignal) ? true : driveSignal.isOpenLoop());
-            // setModuleStates(moduleStates, driveSignal.isOpenLoop());
-        }
-    }
-
-    public void updateLogs() {
-
-        Logger.log("/SwerveDriveSubsystem/Pose", pose);
+    private void logTelemetry(SwerveDriveState state) {
+        Logger.log("/SwerveDriveSubsystem/Pose", state.Pose);
         // Logger.log("/SwerveDriveSubsystem/Velocity", velocity);
         // Logger.log("/SwerveDriveSubsystem/Desired Velocity", (ChassisSpeeds) driveSignal);
 
         Logger.log("/SwerveDriveSubsystem/Velocity Magnitude", getVelocityMagnitude());
 
-        Logger.log("/SwerveDriveSubsystem/Acceleration Commanded", acceleration);
+        Logger.log("/SwerveDriveSubsystem/Acceleration Commanded", 0);
 
-        Logger.log("/SwerveDriveSubsystem/Wheel Speed", modules[0].getState().speedMetersPerSecond);
+        Logger.log("/SwerveDriveSubsystem/Wheel Zero Speed", state.ModuleStates[0].speedMetersPerSecond);
 
-        Logger.log("/SwerveDriveSubsystem/Pitch", getGyroRotation3d().getY());
-        Logger.log("/SwerveDriveSubsystem/Roll", getGyroRotation3d().getX());
-        Logger.log("/SwerveDriveSubsystem/Tilt", getTiltAmountInDegrees());
+        Rotation3d currentRotation = getRotation3d();
+
+        Logger.log("/SwerveDriveSubsystem/Pitch", currentRotation.getY());
+        Logger.log("/SwerveDriveSubsystem/Roll", currentRotation.getX());
+        Logger.log("/SwerveDriveSubsystem/Tilt", getTilt().getRadians());
 
         Logger.log("/SwerveDriveSubsystem/CANCoder Angles", new double[] {
-            modules[0].getCanCoderAngle().getDegrees(),
-            modules[1].getCanCoderAngle().getDegrees(),
-            modules[2].getCanCoderAngle().getDegrees(),
-            modules[3].getCanCoderAngle().getDegrees(),
+            state.ModuleStates[0].angle.getDegrees(),
+            state.ModuleStates[1].angle.getDegrees(),
+            state.ModuleStates[2].angle.getDegrees(),
+            state.ModuleStates[3].angle.getDegrees(),
         });
 
-        // Raw In This Case means the offset is not applied
-        Logger.log("/SwerveDriveSubsystem/Raw CANCoder Angles", new double[] {
-            modules[0].getRawCanCoderAngle().getDegrees(),
-            modules[1].getRawCanCoderAngle().getDegrees(),
-            modules[2].getRawCanCoderAngle().getDegrees(),
-            modules[3].getRawCanCoderAngle().getDegrees(),
-        });
+        // // Raw In This Case means the offset is not applied
+        // Logger.log("/SwerveDriveSubsystem/Raw CANCoder Angles", new double[] {
+        //     state.ModuleStates[0].,
+        //     state.ModuleStates[0].angle.getDegrees(),
+        //     modules[2].getRawCanCoderAngle().getDegrees(),
+        //     modules[3].getRawCanCoderAngle().getDegrees(),
+        // });
 
         Logger.log("/SwerveDriveSubsystem/SwerveModuleStates/Measured", new double[] {
-            modules[0].getState().angle.getRadians(), modules[0].getState().speedMetersPerSecond,
-            modules[1].getState().angle.getRadians(), modules[1].getState().speedMetersPerSecond,
-            modules[2].getState().angle.getRadians(), modules[2].getState().speedMetersPerSecond,
-            modules[3].getState().angle.getRadians(), modules[3].getState().speedMetersPerSecond
+            state.ModuleStates[0].angle.getRadians(), state.ModuleStates[0].speedMetersPerSecond,
+            state.ModuleStates[1].angle.getRadians(), state.ModuleStates[1].speedMetersPerSecond,
+            state.ModuleStates[2].angle.getRadians(), state.ModuleStates[2].speedMetersPerSecond,
+            state.ModuleStates[3].angle.getRadians(), state.ModuleStates[3].speedMetersPerSecond
         });
 
         Logger.log("/SwerveDriveSubsystem/SwerveModuleStates/Setpoints", new double[] {
-            modules[0].getDesiredState().angle.getRadians(), modules[0].getDesiredState().speedMetersPerSecond,
-            modules[1].getDesiredState().angle.getRadians(), modules[1].getDesiredState().speedMetersPerSecond,
-            modules[2].getDesiredState().angle.getRadians(), modules[2].getDesiredState().speedMetersPerSecond,
-            modules[3].getDesiredState().angle.getRadians(), modules[3].getDesiredState().speedMetersPerSecond,
+            state.ModuleTargets[0].angle.getRadians(), state.ModuleTargets[0].speedMetersPerSecond,
+            state.ModuleTargets[1].angle.getRadians(), state.ModuleTargets[1].speedMetersPerSecond,
+            state.ModuleTargets[2].angle.getRadians(), state.ModuleTargets[2].speedMetersPerSecond,
+            state.ModuleTargets[3].angle.getRadians(), state.ModuleTargets[3].speedMetersPerSecond
         });
 
         // Logger.log("/SwerveDriveSubsystem/Wheel Angles", new double[] {
@@ -495,43 +283,17 @@ public class SwerveDriveSubsystem extends CommandSwerveDrivetrain {
         //     modules[3].getAngleVoltage()
         // });
 
-        Logger.log("/SwerveDriveSubsystem/Drive Temperatures", getDriveTemperatures());
-        Logger.log("/SwerveDriveSubsystem/Angle Temperatures", getAngleTemperatures());
-
-        
-    }
-
-    public SwerveModuleState[] getModuleStates() {
-        SwerveModuleState[] states = new SwerveModuleState[4];
-        for (SwerveModule module : modules) {
-            states[module.moduleNumber] = module.getState();
-        }
-        return states;
-    }
-
-    public SwerveModulePosition[] getModulePositions() {
-        SwerveModulePosition[] positions = new SwerveModulePosition[4];
-        for (SwerveModule module : modules) {
-            positions[module.moduleNumber] = module.getPosition();
-        }
-        return positions;
-    }
-
-    public double[] getDriveTemperatures() {
-        return new double[] {
-            modules[0].getDriveTemperature(),
-            modules[1].getDriveTemperature(),
-            modules[2].getDriveTemperature(),
-            modules[3].getDriveTemperature()
-        };
-    }
-
-    public double[] getAngleTemperatures() {
-        return new double[] {
-            modules[0].getAngleTemperature(),
-            modules[1].getAngleTemperature(),
-            modules[2].getAngleTemperature(),
-            modules[3].getAngleTemperature()
-        };
+        Logger.log("/SwerveDriveSubsystem/Drive Temperatures", new double[] {          
+            getModule(0).getDriveMotor().getDeviceTemp().getValue(),
+            getModule(1).getDriveMotor().getDeviceTemp().getValue(),
+            getModule(2).getDriveMotor().getDeviceTemp().getValue(),
+            getModule(3).getDriveMotor().getDeviceTemp().getValue()
+        });
+        Logger.log("/SwerveDriveSubsystem/Angle Temperatures", new double[] {          
+            getModule(0).getSteerMotor().getDeviceTemp().getValue(),
+            getModule(1).getSteerMotor().getDeviceTemp().getValue(),
+            getModule(2).getSteerMotor().getDeviceTemp().getValue(),
+            getModule(3).getSteerMotor().getDeviceTemp().getValue()
+        });
     }
 }
