+++ src/main/java/frc/robot/subsystems/vision/VisionSubsystem.java
@@ -6,19 +6,15 @@ import java.util.Optional;
 
 import org.photonvision.targeting.PhotonTrackedTarget;
 
-import frc.lib.logging.Logger;
-
 import edu.wpi.first.math.Matrix;
 import edu.wpi.first.math.VecBuilder;
 import edu.wpi.first.math.geometry.Pose3d;
 import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
-import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.lib.math.MathUtils;
 import frc.lib.vision.LimelightRawAngles;
 import frc.robot.Constants.FieldConstants;
-import frc.robot.subsystems.swervedrive.CommandSwerveDrivetrain;
 import frc.robot.subsystems.swervedrive.SwerveDriveSubsystem;
 import frc.robot.subsystems.vision.AprilTagIO.AprilTagIOInputs;
 import frc.robot.subsystems.vision.PositionTargetIO.PositionTargetIOInputs;
@@ -38,9 +34,9 @@ public class VisionSubsystem extends SubsystemBase {
     public List<PhotonTrackedTarget> leftTargets = new ArrayList<>();
     public List<PhotonTrackedTarget> rightTargets = new ArrayList<>();
 
-    private CommandSwerveDrivetrain consumer;
+    private SwerveDriveSubsystem consumer;
 
-    public VisionSubsystem(CommandSwerveDrivetrain consumer, AprilTagIO left, AprilTagIO right, PositionTargetIO limelight) {
+    public VisionSubsystem(SwerveDriveSubsystem consumer, AprilTagIO left, AprilTagIO right, PositionTargetIO limelight) {
         this.left = left;
         this.right = right;
         this.limelight = limelight;
